refactor(create-agentuity): convert to TypeScript package
create-agentuity was the only package with bin.js / bin.test.js at the
root and a no-op build, which broke scripts/build-sdk.sh's dist/
verification. Converting it to a standard TS package brings it in line
with every other package and lets tsc --build handle it.

Structure:
- src/index.ts    — exports getCliVersionSpecifier
- src/bin.ts      — the #!/usr/bin/env node entry (spawns bunx
                    @agentuity/cli@<tag>)
- test/index.test.ts — same 14 tests, now imports from src/
- tsconfig.json   — standard shape (extends tsconfig.base.json)

package.json changes:
- bin: ./bin.js -> ./dist/bin.js
- adds main + types + exports fields pointing at dist/
- files: ["bin.js"] -> ["dist", "README.md"]
- scripts: no-op build -> real bunx tsc --build; adds clean, typecheck,
  prepublishOnly matching other packages
- devDeps: adds @types/bun, typescript

Runtime note: dist/bin.js reads its own package.json via
readFileSync(resolve(__dirname, '..', 'package.json'), 'utf8') instead
of createRequire, which is simpler and Node-native. Users installing
create-agentuity via 'bun create agentuity' still get the same behaviour
— package.json sits one level up from dist/ in the published tarball.

Root tsconfig.json adds packages/create-agentuity to its references so
'bunx tsc --build' picks it up.

Verification:
- bun install                     ✓ (bun.lock updated)
- bun run build                   ✓
- bun run typecheck               ✓ 29 packages + 4 apps clean
- bash scripts/build-sdk.sh       ✓ now lists 'All 28 packages built
                                    successfully' (create-agentuity
                                    now produces dist/)
- cd packages/create-agentuity && bun test  ✓ 14/14 pass
